feat: current time millis as report version.

+++ accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/domain/entity/report/ReportEntity.java
@@ -53,7 +53,7 @@ public class ReportEntity extends CommonEntity implements Persistable<String>, V
     @Column(name = "ver", nullable = false)
     @Getter
     @Setter
-    private Integer ver = 1;
+    private long ver = 1;
 
     @Override
     public String getId() {
